+++ src/Application.cpp
@@ -179,6 +179,27 @@ int Application::run(QApplication &qtApp)
         this->windows->forceLayoutChannelViews();
     });
 
+    getSettings()->enableBTTVGlobalEmotes.connect(
+        [this] {
+            this->twitch->reloadBTTVGlobalEmotes();
+        },
+        false);
+    getSettings()->enableBTTVChannelEmotes.connect(
+        [this] {
+            this->twitch->reloadAllBTTVChannelEmotes();
+        },
+        false);
+    getSettings()->enableFFZGlobalEmotes.connect(
+        [this] {
+            this->twitch->reloadFFZGlobalEmotes();
+        },
+        false);
+    getSettings()->enableFFZChannelEmotes.connect(
+        [this] {
+            this->twitch->reloadAllFFZChannelEmotes();
+        },
+        false);
+
     return qtApp.exec();
 }