+++ src/common/Channel.hpp
@@ -57,8 +57,12 @@ public:
         messageAppended;
     pajlada::Signals::Signal<std::vector<MessagePtr> &> messagesAddedAtStart;
     pajlada::Signals::Signal<size_t, MessagePtr &> messageReplaced;
+    /// Invoked when some number of messages were filled in using time received
+    pajlada::Signals::Signal<const std::vector<MessagePtr> &> filledInMessages;
     pajlada::Signals::NoArgSignal destroyed;
     pajlada::Signals::NoArgSignal displayNameChanged;
+    /// Invoked when AbstractIrcServer::onReadConnected occurs
+    pajlada::Signals::NoArgSignal connected;
 
     Type getType() const;
     const QString &getName() const;
@@ -75,12 +79,17 @@ public:
     void addMessage(
         MessagePtr message,
         boost::optional<MessageFlags> overridingFlags = boost::none);
-    void addMessagesAtStart(std::vector<MessagePtr> &messages_);
+    void addMessagesAtStart(const std::vector<MessagePtr> &messages_);
+
+    /// Inserts the given messages in order by Message::serverReceivedTime.
+    void fillInMissingMessages(const std::vector<MessagePtr> &messages);
+
     void addOrReplaceTimeout(MessagePtr message);
     void disableAllMessages();
     void replaceMessage(MessagePtr message, MessagePtr replacement);
     void replaceMessage(size_t index, MessagePtr replacement);
     void deleteMessage(QString messageID);
+
     MessagePtr findMessage(QString messageID);
 
     bool hasMessages() const;